+++ tests/test_config.py
@@ -6,13 +6,17 @@ CURR_DIR=os.path.dirname(os.path.realpath(__file__))
 
 class TestConfig(unittest.TestCase):
 
-    def testShouldRaiseErrorOnUnexistingFile(self):
+    #
+    # read()
+    #
+
+    def testReadShouldRaiseErrorOnUnexistingFile(self):
         config = Config()
 
         with self.assertRaises(Exception):
             config.read("/path/to/none.yml")
 
-    def testShouldSupportAnEmptyConfigFile(self):
+    def testReadShouldSupportAnEmptyConfigFile(self):
         config = Config()
         config.read(CURR_DIR + "/fixtures/config-empty.yml")
 
@@ -20,7 +24,7 @@ class TestConfig(unittest.TestCase):
         self.assertEqual(config.getExporterPort(), 9100)
         self.assertEqual(config.getPgbouncers(), [])
 
-    def testShouldParseConfigFileWithOnePgbouncer(self):
+    def testReadShouldParseConfigFileWithOnePgbouncer(self):
         config = Config()
         config.read(CURR_DIR + "/fixtures/config-with-one-pgbouncer.yml")
 
@@ -35,7 +39,7 @@ class TestConfig(unittest.TestCase):
         self.assertEqual(config.getPgbouncers()[0].getExcludeDatabases(), ["three"])
         self.assertEqual(config.getPgbouncers()[0].getExtraLabels(), {"first": "1", "second": "2"})
 
-    def testShouldParseConfigFileWithTwoPgbouncer(self):
+    def testReadShouldParseConfigFileWithTwoPgbouncer(self):
         config = Config()
         config.read(CURR_DIR + "/fixtures/config-with-two-pgbouncer.yml")
 
@@ -57,7 +61,7 @@ class TestConfig(unittest.TestCase):
         self.assertEqual(config.getPgbouncers()[1].getExcludeDatabases(), [])
         self.assertEqual(config.getPgbouncers()[1].getExtraLabels(), {})
 
-    def testShouldInjectEnvironmentVariablesOnParsing(self):
+    def testReadShouldInjectEnvironmentVariablesOnParsing(self):
         os.environ["TEST_USERNAME"] = "marco"
         os.environ["TEST_PASSWORD"] = "secret"
         os.environ["TEST_INCLUDE_DATABASE"] = "production"
@@ -72,7 +76,7 @@ class TestConfig(unittest.TestCase):
         self.assertEqual(config.getPgbouncers()[0].getIncludeDatabases(), ["production"])
         self.assertEqual(config.getPgbouncers()[0].getExtraLabels(), {"cluster": "users-1-1000"})
 
-    def testShouldInjectEnvironmentVariablesOnParsingEvenIfEmpty(self):
+    def testReadShouldInjectEnvironmentVariablesOnParsingEvenIfEmpty(self):
         os.environ["TEST_USERNAME"] = "marco"
         os.environ["TEST_PASSWORD"] = ""
         os.environ["TEST_INCLUDE_DATABASE"] = "production"
@@ -85,7 +89,7 @@ class TestConfig(unittest.TestCase):
         self.assertEqual(len(config.getPgbouncers()), 1)
         self.assertEqual(config.getPgbouncers()[0].getDsn(), "postgresql://marco:@host:6431/pgbouncer")
 
-    def testShouldKeepOriginalConfigOnMissingEnvironmentVariables(self):
+    def testReadShouldKeepOriginalConfigOnMissingEnvironmentVariables(self):
         del os.environ["TEST_USERNAME"]
         os.environ["TEST_PASSWORD"] = "secret"
         del os.environ["TEST_INCLUDE_DATABASE"]
@@ -99,6 +103,65 @@ class TestConfig(unittest.TestCase):
         self.assertEqual(config.getPgbouncers()[0].getDsn(), "postgresql://$(TEST_USERNAME):secret@host:6431/pgbouncer")
         self.assertEqual(config.getPgbouncers()[0].getIncludeDatabases(), ["$(TEST_INCLUDE_DATABASE)"])
         self.assertEqual(config.getPgbouncers()[0].getExtraLabels(), {"cluster": "users-1-1000"})
+
+    #
+    # validate()
+    #
+
+    def testValidateShouldPassOnConfigContainingOnePgbouncer(self):
+        config = Config()
+        config.read(CURR_DIR + "/fixtures/config-with-one-pgbouncer.yml")
+        config.validate()
+
+    def testValidateShouldPassOnConfigContainingTwoPgbouncer(self):
+        config = Config()
+        config.read(CURR_DIR + "/fixtures/config-with-two-pgbouncer.yml")
+        config.validate()
+
+    def testValidateShouldRaiseExceptionOnNoPgbouncerConfigured(self):
+        config = Config({})
+
+        with self.assertRaisesRegex(Exception, "no pgbouncer instance configured"):
+            config.validate()
+
+    def testValidateShouldRaiseExceptionOnTwoPgbouncersWithNoLabels(self):
+        config = Config({"pgbouncers": [
+            {"dsn": "postgresql://"},
+            {"dsn": "postgresql://"}
+        ]})
+
+        with self.assertRaisesRegex(Exception, "extra_labels configured for each pgbouncer must be unique"):
+            config.validate()
+
+    def testValidateShouldRaiseExceptionOnTwoPgbouncersWithSameExtraLabels(self):
+        config = Config({"pgbouncers": [
+            {"dsn": "postgresql://", "extra_labels": {"pool_id": 1, "another_id": 5}},
+            {"dsn": "postgresql://", "extra_labels": {"pool_id": 1, "another_id": 5}}
+        ]})
+
+        with self.assertRaisesRegex(Exception, "extra_labels configured for each pgbouncer must be unique"):
+            config.validate()
+
+    def testValidateShouldPassOnTwoPgbouncersWithDifferentExtraLabels(self):
+        config = Config({"pgbouncers": [
+            {"dsn": "postgresql://", "extra_labels": {"pool_id": 1, "another_id": 5}},
+            {"dsn": "postgresql://", "extra_labels": {"pool_id": 2, "another_id": 5}}
+        ]})
+
+        config.validate()
+
+
+class TestPgbouncerConfig(unittest.TestCase):
+
+    def testValidateShouldPassOnConfigContainingOnlyDsn(self):
+        config = PgbouncerConfig({"dsn": "postgresql://"})
+        config.validate()
+
+    def testValidateShouldRaiseExceptionOnEmptyDsn(self):
+        config = PgbouncerConfig({"dsn": ""})
+
+        with self.assertRaisesRegex(Exception, "The DSN is required"):
+            config.validate()
 
 
 if __name__ == '__main__':